refactor(api): change GroupUse class to a static inner class of Group
